Add roving-packing-list
This file lists everything we should have for going roving, who is responsible for it, and a column for checking off if it's packed (like if you print this file out).  As things are purchased or built, my intention is that they move from vhf-contest-improvements to this file.  Consider that file a TODO and this file a list of things we need to gather up.